chore: re-organize folder paths in a React child components pattern
