+++ src/App.css
@@ -1,7 +1,35 @@
-* {
+/* * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+} */
+
+/* Reset styles for body and html */
+html,
+body {
+  margin: 0;
+  padding: 0;
+  overflow-x: hidden; /* Prevent horizontal scroll */
+}
+
+/* Ensure #root and its parent elements take full width and height */
+#root,
+.main,
+.background {
+  width: 100%;
+  margin: 0;
+  padding: 0;
+  display: flex; /* Or block depending on your layout needs */
+}
+
+/* Ensure background covers the entire area */
+.background {
+  background: url("./assets/scale.webp") no-repeat center center fixed;
+  background-size: cover;
+  width: 100%;
+  height: 100%;
+  min-width: 100vw;
+  min-height: 100vh;
 }
 
 @import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
@@ -41,8 +69,8 @@ img {
 }
 
 .main {
-  width: 100%;
-  max-width: 100vw;
+  /* width: 100%; */
+  /* max-width: 100vw; */
   height: 100%;
   position: relative;
   background-color: var(--primary-color);
@@ -59,29 +87,18 @@ img {
   filter: brightness(85%);
 } */
 
-.background {
-  position: absolute;
-  height: 100vh;
-  width: 100%;
-  background: url("./assets/background-large.avif");
-  background-position: center;
-  background-size: cover;
-  filter: brightness(80%);
-  z-index: 0;
-}
-
 .navigation {
-  border-radius: 12px;
-  height: 80px;
+  border-radius: 5%;
+  height: 40px;
   position: fixed;
   display: flex;
-  align-items: center;
+  align-items: normal;
   justify-content: center;
   background-color: rgba(0, 0, 0, 0.4);
   -webkit-backdrop-filter: blur(10px);
   backdrop-filter: blur(10px);
-  padding: 2vh 0 2vh 0;
-  margin: 25px 0 10px 0;
+  padding: 1vh 1vw;
+  margin: 25px 0 0px 0;
   left: 50%;
   transform: translate(-50%);
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
@@ -99,23 +116,28 @@ img {
   white-space: nowrap;
   text-align: center;
   color: var(--quarternary-color);
-  font-size: 30px;
-  height: 30px;
+  font-size: 1.2em;
+  height: 20px;
   background-color: transparent;
   background-repeat: no-repeat;
   border: 0;
-  margin: 0 10px 0 10px;
+  margin: 0 15px 0 10px;
   flex-basis: auto;
-  border-radius: 10%;
   box-shadow: inset 0 0 0 0 rgba(211, 122, 71, 0.8);
   transition: linear 0.8s;
+}
+
+.separator {
+  margin: 0 10px;
+  color: white;
+  font-size: 1.2em;
 }
 
 .navigation-button:hover {
   /* background-color: transparent; */
   /* color: var(--secondary-color);
     transform: scale(1.1); */
-  box-shadow: inset 20vw 0 0 0 rgba(211, 122, 71, 0.8);
+  /* box-shadow: inset 70vw 40vw 40vw 0 rgba(211, 122, 71, 0.8); */
   cursor: pointer;
 }
 
@@ -547,9 +569,19 @@ img {
   }
 
   .background {
-    background: url("./assets/scale.webp");
-    background-position: center;
+    background: url("./assets/scale.webp") no-repeat center center fixed;
     background-size: cover;
+    width: 100vw;
+    height: 100vh;
+  }
+
+  .container,
+  .wrapper,
+  .parent-class-of-background {
+    width: 100%;
+    height: 100%;
+    margin: 0;
+    padding: 0;
   }
 
   .navigation {
