--- src/TextFormatting/StylizedParagraphs.css
@@ -1,18 +1,0 @@
-.first-character {
-  color: var(-color);
-  float: left;
-  font-family: Georgia;
-  font-size: 6vw;
-  line-height: 60px;
-  padding-top: 4px;
-  padding-right: 8px;
-  padding-left: 3px;
-}
-
-.paragraph {
-  margin-left: 4%;
-  width: 40%;
-  line-height: 190%;
-  display: inline-block;
-  min-width: 375px;
-}
