+++ src/components/BuyAHome/BuyAHome.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
 import Modal from "react-modal";
 import "./BuyAHome.css";
-import config from "../config.js";
+import config from "../../config.js";
 import ListingsLayout from "./Listings/ListingsLayout.js";
-import beforeYouBuy from "../assets/buy-a-home/illustrations/update1-03.avif";
-import findingTheDeal from "../assets/buy-a-home/illustrations/update2-02.avif";
-import sealingTheDeal from "../assets/buy-a-home/illustrations/update3-03.avif";
-import closeModalImg from "../assets/close-modal.avif";
+import beforeYouBuy from "../../assets/buy-a-home/illustrations/update1-03.avif";
+import findingTheDeal from "../../assets/buy-a-home/illustrations/update2-02.avif";
+import sealingTheDeal from "../../assets/buy-a-home/illustrations/update3-03.avif";
+import closeModalImg from "../../assets/close-modal.avif";
 
 Modal.setAppElement("#root");
