+++ src/components/BuyAHome/Listings/ListingsLayout.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./ListingsLayout.css";
-import config from "../../config.js";
+import config from "../../../config.js";
 import Listing from "./Listing.js";
 
 function ListingsLayout({ listings }) {
